+++ packages/expo-sqlite/ios/SQLiteOptions.swift
@@ -15,12 +15,28 @@ struct OpenDatabaseOptions: Record, Equatable {
   @Field
   var finalizeUnusedStatementsBeforeClosing: Bool = true
 
+  @Field
+  var libSQLUrl: URL?
+
+  @Field
+  var libSQLAuthToken: String?
+
+  @Field
+  var libSQLRemoteOnly: Bool = false
+
+  @Field
+  var libSQLSyncInterval: Int = 0
+
   // MARK: - Equatable
 
   static func == (lhs: OpenDatabaseOptions, rhs: OpenDatabaseOptions) -> Bool {
     return lhs.enableCRSQLite == rhs.enableCRSQLite &&
       lhs.enableChangeListener == rhs.enableChangeListener &&
       lhs.useNewConnection == rhs.useNewConnection &&
-      lhs.finalizeUnusedStatementsBeforeClosing == rhs.finalizeUnusedStatementsBeforeClosing
+      lhs.finalizeUnusedStatementsBeforeClosing == rhs.finalizeUnusedStatementsBeforeClosing &&
+      lhs.libSQLUrl == rhs.libSQLUrl &&
+      lhs.libSQLAuthToken == rhs.libSQLAuthToken &&
+      lhs.libSQLRemoteOnly == rhs.libSQLRemoteOnly &&
+      lhs.libSQLSyncInterval == rhs.libSQLSyncInterval
   }
 }